feat(dashboard): v1-03 visual rewrite — decompose, tokenize, sidebar layout
- Decompose DashboardPage.tsx from 929 → ~280 lines (AC-12)
- Extract: DashboardHero, DashboardProjectGrid, DashboardSidebar,
  DashboardEmptyState, useDashboardActions, useDashboardLayout,
  dashboardUtils
- Tokenize HeroCard: all arbitrary values → Design Token refs (AC-2)
- Replace native <button> → Button primitive (AC-3)
- Add card hover effects with shadow tokens (AC-4)
- Add right sidebar with recent docs, quick actions, stats (AC-1, AC-5)
- Enhanced empty state per 26-empty-states.html (AC-6)
- All new code uses semantic Design Tokens (AC-7)
- Storybook builds clean (AC-8)
- All 2051 tests pass across 288 files (AC-9)
- TypeScript strict passes (AC-10)
- ESLint 0 errors (AC-11)

Closes #1167
