+++ apps/desktop/renderer/src/features/dashboard/useDashboardLayout.ts
@@ -1,0 +1,17 @@
+import React from "react";
+
+/**
+ * Hook managing dashboard layout state — sidebar visibility and responsive breakpoints.
+ */
+export function useDashboardLayout() {
+  const [sidebarVisible, setSidebarVisible] = React.useState(true);
+
+  const toggleSidebar = React.useCallback(() => {
+    setSidebarVisible((prev) => !prev);
+  }, []);
+
+  return {
+    sidebarVisible,
+    toggleSidebar,
+  };
+}
